Require JDK1.7 to compile
But generate code that can run on JDK6. In this way we
can compile JDK7-specific code: as long as it is only
invoked when running on JDK7, and we provide a JDK6
specific alternative, we are good. As long as we do
not initialize the JDK7-specific classes, in theory
we should be good.